uint int comparison fix in rendermodel.cpp

+++ src/engine/render/rendermodel.cpp
@@ -1100,7 +1100,7 @@ void startmodelquery(occludequery *query)
 
 void endmodelquery()
 {
-    if(batchedmodels.size() == modelquerymodels)
+    if(static_cast<int>(batchedmodels.size()) == modelquerymodels)
     {
         modelquery->fragments = 0;
         modelquery = nullptr;
